# Optimize exporters tests