--- a/exporters/elasticsearch-exporter/src/test/java/io/camunda/zeebe/exporter/ElasticsearchExporterIT.java
+++ b/exporters/elasticsearch-exporter/src/test/java/io/camunda/zeebe/exporter/ElasticsearchExporterIT.java
@@ -26,10 +26,12 @@ import java.util.Collections;
 import java.util.Map;
 import java.util.UUID;
 import org.agrona.CloseHelper;
-import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assumptions;
-import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.testcontainers.elasticsearch.ElasticsearchContainer;
@@ -44,6 +46,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
  * down, should be done elsewhere (e.g. {@link FaultToleranceIT}
  */
 @Testcontainers
+@TestInstance(Lifecycle.PER_CLASS)
 final class ElasticsearchExporterIT {
   @Container
   private static final ElasticsearchContainer CONTAINER = TestSupport.createDefaultContainer();
@@ -57,8 +60,8 @@ final class ElasticsearchExporterIT {
 
   private TestClient testClient;
 
-  @BeforeEach
-  public void beforeEach() {
+  @BeforeAll
+  public void beforeAll() {
     // as all tests use the same endpoint, we need a per-test unique prefix
     config.index.prefix = UUID.randomUUID() + "-test-record";
     config.url = CONTAINER.getHttpHostAddress();
@@ -66,6 +69,9 @@ final class ElasticsearchExporterIT {
     config.index.setNumberOfReplicas(1);
     config.index.createTemplate = true;
     config.bulk.size = 1; // force flushing on the first record
+    // here; enable all indexes that needed during the tests beforehand as they will be created once
+    config.index.jobBatch = true;
+
     testClient = new TestClient(config, indexRouter);
 
     exporter.configure(
@@ -74,8 +80,8 @@ final class ElasticsearchExporterIT {
     exporter.open(controller);
   }
 
-  @AfterEach
-  void afterEach() {
+  @AfterAll
+  void afterAll() {
     CloseHelper.quietCloseAll(testClient);
   }
 
@@ -84,7 +90,6 @@ final class ElasticsearchExporterIT {
   void shouldExportRecord(final ValueType valueType) {
     // given
     final var record = factory.generateRecord(valueType);
-    TestSupport.setIndexingForValueType(config.index, valueType, true);
 
     // when
     exporter.export(record);
@@ -115,7 +120,6 @@ final class ElasticsearchExporterIT {
             .build();
     final Record<JobRecordValue> record =
         factory.generateRecord(ValueType.JOB, builder -> builder.withValue(value));
-    config.index.job = true;
 
     // when
     exporter.export(record);
@@ -142,7 +146,6 @@ final class ElasticsearchExporterIT {
             .build();
     final Record<JobBatchRecordValue> record =
         factory.generateRecord(ValueType.JOB_BATCH, builder -> builder.withValue(value));
-    config.index.jobBatch = true;
 
     // when
     exporter.export(record);
@@ -163,7 +166,6 @@ final class ElasticsearchExporterIT {
     // given
     final var record = factory.generateRecord(valueType);
     final var expectedIndexTemplateName = indexRouter.indexPrefixForValueType(valueType);
-    TestSupport.setIndexingForValueType(config.index, valueType, true);
 
     // when - export a single record to enforce installing all index templatesWrapper
     exporter.export(record);