--- a/exporters/opensearch-exporter/src/test/java/io/camunda/zeebe/exporter/opensearch/OpensearchExporterIT.java
+++ b/exporters/opensearch-exporter/src/test/java/io/camunda/zeebe/exporter/opensearch/OpensearchExporterIT.java
@@ -25,10 +25,12 @@ import java.util.Collections;
 import java.util.Map;
 import java.util.UUID;
 import org.agrona.CloseHelper;
-import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assumptions;
-import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.opensearch.client.opensearch.core.GetResponse;
@@ -44,6 +46,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
  * down, should be done elsewhere (e.g. {@link FaultToleranceIT}
  */
 @Testcontainers
+@TestInstance(Lifecycle.PER_CLASS)
 final class OpensearchExporterIT {
   @Container
   private static final OpensearchContainer CONTAINER = TestSupport.createDefaultContainer();
@@ -56,8 +59,8 @@ final class OpensearchExporterIT {
 
   private TestClient testClient;
 
-  @BeforeEach
-  public void beforeEach() {
+  @BeforeAll
+  public void beforeAll() {
     // as all tests use the same endpoint, we need a per-test unique prefix
     config.index.prefix = UUID.randomUUID() + "-test-record";
     config.url = CONTAINER.getHttpHostAddress();
@@ -65,6 +68,9 @@ final class OpensearchExporterIT {
     config.index.setNumberOfReplicas(1);
     config.index.createTemplate = true;
     config.bulk.size = 1; // force flushing on the first record
+    // here; enable all indexes that needed during the tests beforehand as they will be created once
+    config.index.jobBatch = true;
+
     testClient = new TestClient(config, indexRouter);
 
     exporter.configure(
@@ -73,8 +79,8 @@ final class OpensearchExporterIT {
     exporter.open(controller);
   }
 
-  @AfterEach
-  void afterEach() {
+  @AfterAll
+  void afterAll() {
     CloseHelper.quietCloseAll(testClient);
   }
 
@@ -83,7 +89,6 @@ final class OpensearchExporterIT {
   void shouldExportRecord(final ValueType valueType) {
     // given
     final var record = factory.generateRecord(valueType);
-    TestSupport.setIndexingForValueType(config.index, valueType, true);
 
     // when
     exporter.export(record);
@@ -114,7 +119,6 @@ final class OpensearchExporterIT {
             .build();
     final Record<JobRecordValue> record =
         factory.generateRecord(ValueType.JOB, builder -> builder.withValue(value));
-    config.index.job = true;
 
     // when
     exporter.export(record);
@@ -141,7 +145,6 @@ final class OpensearchExporterIT {
             .build();
     final Record<JobBatchRecordValue> record =
         factory.generateRecord(ValueType.JOB_BATCH, builder -> builder.withValue(value));
-    config.index.jobBatch = true;
 
     // when
     exporter.export(record);
@@ -162,7 +165,6 @@ final class OpensearchExporterIT {
     // given
     final var record = factory.generateRecord(valueType);
     final var expectedIndexTemplateName = indexRouter.indexPrefixForValueType(valueType);
-    TestSupport.setIndexingForValueType(config.index, valueType, true);
 
     // when - export a single record to enforce installing all index templatesWrapper
     exporter.export(record);