@abbasadel 2023-08-18
> Do you know if this is differs from using a `beforeAll` and making all fields static?

In most cases, `beforeAll` works similarly to the static initialization. I avoided using static block since it was conflicting with the test container's initialization.